+++ src/modules/v1/farmer/farmer-model.ts
@@ -17,7 +17,6 @@ const farmerSchema = new Schema<IFarmer>(
     name: {
       type: String,
       required: true,
-      index: true,
       trim: true,
     },